+++ src/debug/mod.rs
@@ -378,8 +378,8 @@ pub fn debug_draw(
                 ),
                 style: Style {
                     position_type: PositionType::Absolute,
-                    left: Val::Px(location.position.x + 5.0) / scale.0 as f32,
-                    top: Val::Px(location.position.y + 5.0) / scale.0 as f32,
+                    left: Val::Px(location.position.x + 5.0) / scale.0,
+                    top: Val::Px(location.position.y + 5.0) / scale.0,
                     ..Default::default()
                 },
                 ..Default::default()
